refactor: Migrate payment template in a Gradle module.
