+++ app/src/main/java/com/guru/composecookbook/ui/templates/TemplatesActivity.kt
@@ -14,7 +14,7 @@ import com.guru.composecookbook.comingsoon.ComingSoon
 import com.guru.composecookbook.theme.ComposeCookBookTheme
 import com.guru.composecookbook.ui.templates.logins.LoginOnboarding
 import com.guru.composecookbook.onboarding.OnBoardingScreen
-import com.guru.composecookbook.ui.templates.payment.AddPaymentCard
+import com.guru.composecookbook.paymentcard.AddPaymentScreen
 import com.guru.composecookbook.ui.templates.profile.ProfileScreen
 
 class TemplatesActivity : ComponentActivity() {
@@ -55,7 +55,7 @@ fun TemplateApp(templateType: String) {
         "Login" -> LoginOnboarding()
         "On-boarding" -> OnBoardingScreen { }
         "Charts" -> Charts()
-        "Adding Payment Card" -> AddPaymentCard()
+        "Adding Payment Card" -> AddPaymentScreen()
         else -> ComingSoon()
     }
 }
