+++ app/src/main/res/values/strings.xml
@@ -23,10 +23,4 @@
     <string name="title_activity_custom_fling">Custom Fling</string>
     <string name="title_activity_custom_fling_settings">Custom Fling Settings</string>
     <string name="title_activity_templates">Templates</string>
-    <string name="card_holder_name">Card holder name</string>
-    <string name="card_holder_number">Card number</string>
-    <string name="expiry_date">Expiry date</string>
-    <string name="cvc">CVC</string>
-    <string name="save">save</string>
-
 </resources>
