+++ settings.gradle.kts
@@ -24,3 +24,4 @@ include(":animations:lottie")
 include(":animations:canvas")
 include(":components:comingsoon")
 include(":templates:onboarding")
+include(":templates:paymentcard")
