+++ templates/paymentcard/src/main/res/drawable/card_symbol.png
